Move Vulkan Version to DeviceInfo + hardcode to VK12
